EZP-29812: As an editor, I want to embed a content item within a block element
