+++ src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-paragraph.js
@@ -9,6 +9,7 @@ export default class EzParagraphConfig extends EzConfigBase {
             'ezmoveup',
             'ezmovedown',
             this.getStyles(config.customStyles),
+            'ezembedinline',
             'ezblocktextalignleft',
             'ezblocktextaligncenter',
             'ezblocktextalignright',
@@ -34,6 +35,6 @@ export default class EzParagraphConfig extends EzConfigBase {
         const nativeEditor = payload.editor.get('nativeEditor');
         const path = nativeEditor.elementPath();
 
-        return (nativeEditor.isSelectionEmpty() && path && path.contains('p'));
+        return nativeEditor.isSelectionEmpty() && path && path.contains('p');
     }
 }
