+++ webpack.config.js
@@ -35,6 +35,7 @@ module.exports = {
         ezBtnImageLink: './src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-imagelink.js',
         ezBtnImageLinkEdit: './src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-imagelinkedit.js',
         ezBtnEmbed: './src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-embed.js',
+        ezBtnEmbedInline: './src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-embedinline.js',
         ezBtnEmbedUpdate: './src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-embedupdate.js',
         ezBtnEmbedAlingCenter: './src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-embedaligncenter.js',
         ezBtnEmbedAlingLeft: './src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-embedalignleft.js',
@@ -53,10 +54,12 @@ module.exports = {
         ezParagraphConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-paragraph.js',
         ezFormattedConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-formatted.js',
         ezTextConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-text.js',
+        ezListConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-list.js',
         ezTableConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-table.js',
         ezLinkConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-link.js',
         ezHeadingConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-heading.js',
         ezEmbedConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-embed.js',
+        ezEmbedInlineConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-embed-inline.js',
         ezEmbedImageConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-image.js',
         ezEmbedImageLinkConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-image-link.js',
         ezCustomTagConfig: './src/bundle/Resources/public/js/alloyeditor/src/toolbars/config/ez-custom-tag.js',
